MR: feature/diary-list -> dev

See merge request sw_track/class_05/web_2_project/team07/front-end!63

+++ src/pages/diaryPage/DiaryPage.tsx
@@ -4,7 +4,7 @@ import ListView from './ListView.tsx';
 import GalleryView from './GalleryView.tsx';
 import Header from '@/components/header/Header';
 import Footer from '@/components/footer/Footer';
-import { BsList, BsFillGridFill } from 'react-icons/bs';
+import { TabImages } from '@/constants/diary';
 import './diaryPage.scss';
 import { db } from '@/firebaseApp.ts';
 import {
@@ -16,6 +16,7 @@ import {
   doc,
   Timestamp,
 } from 'firebase/firestore';
+import { useAuth } from '@/hooks';
 
 interface DiaryProps {
   userEmail: string;
@@ -26,51 +27,48 @@ interface DiaryProps {
   imgUrls: string[];
 }
 
-const Tab = ({ icon, tabName, currentTab, handleTabChange }) => (
-  <div
-    className={`view_tab ${tabName} ${currentTab === tabName ? 'on' : ''}`}
-    onClick={() => handleTabChange(tabName)}
-  >
-    {icon}
-  </div>
-);
-
 const DiaryPage = () => {
+  const user = useAuth();
   const [currentTab, setCurrentTab] = useState('list_tab');
   const [diaryData, setDiaryData] = useState<DiaryProps[]>([]);
 
   useEffect(() => {
     const fetchData = async () => {
-      const userEmail = 'test@test.com';
-      const q = query(
-        collection(db, 'diary'),
-        where('userEmail', '==', userEmail),
-      );
-      const querySnapshot = await getDocs(q);
-      const data: DiaryProps[] = [];
-      querySnapshot.forEach(doc => {
-        data.push({ id: doc.id, ...doc.data() });
-      });
+      if (user) {
+        const q = query(
+          collection(db, 'diary'),
+          where('userEmail', '==', user?.email),
+        );
+        const querySnapshot = await getDocs(q);
+        const data: DiaryProps[] = [];
+        querySnapshot.forEach(doc => {
+          data.push({ id: doc.id, ...doc.data() });
+        });
 
-      const sortedData = data.sort(
-        (a, b) => b.postedAt.toDate() - a.postedAt.toDate(),
-      );
-      setDiaryData(sortedData);
+        const sortedData = data.sort(
+          (a, b) => b.postedAt.toDate() - a.postedAt.toDate(),
+        );
+        setDiaryData(sortedData);
+      }
     };
 
     fetchData();
-  }, []);
+  }, [user]);
 
   const handleDelete = async (index: number) => {
-    try {
-      const diaryIdToDelete = diaryData[index].id;
+    const confirmed = window.confirm('글을 삭제하시겠습니까?');
 
-      await deleteDoc(doc(db, 'diary', diaryIdToDelete));
+    if (confirmed) {
+      try {
+        const diaryIdToDelete = diaryData[index].id;
 
-      const updatedDiaryData = diaryData.filter((_, i) => i !== index);
-      setDiaryData(updatedDiaryData);
-    } catch (error) {
-      console.error('일기 삭제 중 오류:', error);
+        await deleteDoc(doc(db, 'diary', diaryIdToDelete));
+
+        const updatedDiaryData = diaryData.filter((_, i) => i !== index);
+        setDiaryData(updatedDiaryData);
+      } catch (error) {
+        console.error('일기 삭제 중 오류:', error);
+      }
     }
   };
 
@@ -80,44 +78,63 @@ const DiaryPage = () => {
     }
   };
 
-  const tabs = [
-    { icon: <BsList />, tabName: 'list_tab' },
-    { icon: <BsFillGridFill />, tabName: 'gallery_tab' },
+  const tabData = [
+    {
+      name: 'list_tab',
+      label: 'List',
+      onImage: TabImages.LISTON,
+      offImage: TabImages.LISTOFF,
+    },
+    {
+      name: 'gallery_tab',
+      label: 'Gallery',
+      onImage: TabImages.GALLERYON,
+      offImage: TabImages.GALLERYOFF,
+    },
   ];
 
   return (
-    <main className="diary_page">
-      <div className="diary_container">
-        <Header />
-        <h2 className="title inner">
-          <span>{'Joy'}</span>님, 식물의 성장 기록을 남겨보세요
-          <span className="plant_icon"></span>
-        </h2>
-        <section className="view_section">
-          {tabs.map((tab, index) => (
-            <Tab
-              key={index}
-              icon={tab.icon}
-              tabName={tab.tabName}
-              currentTab={currentTab}
-              handleTabChange={handleTabChange}
-            />
-          ))}
-        </section>
-        <section className="content_section">
-          {currentTab === 'list_tab' ? (
-            <ListView diaryData={diaryData} handleDelete={handleDelete} />
-          ) : (
-            <GalleryView diaryData={diaryData} />
-          )}
-        </section>
-        <div className="top_btn"></div>
-        <Footer />
-      </div>
-      <div className="write_btn_wrap">
-        <Link to="/diary/write" className="write_btn"></Link>
-      </div>
-    </main>
+    <>
+      <Header />
+      <main className="diary_page">
+        <div className="diary_container">
+          <h2 className="title inner">
+            <span>{user?.displayName ?? '사용자'}</span>님, 식물의 성장 기록을
+            남겨보세요
+            <span className="plant_icon"></span>
+          </h2>
+          <section className="view_section">
+            {tabData.map((tab, index) => (
+              <div
+                key={index}
+                className={`view_tab ${tab.name} ${
+                  currentTab === tab.name ? 'on' : ''
+                }`}
+                onClick={() => handleTabChange(tab.name)}
+              >
+                <img
+                  src={currentTab === tab.name ? tab.onImage : tab.offImage}
+                  className="tab_img"
+                  alt={`Tab ${tab.label}`}
+                ></img>
+              </div>
+            ))}
+          </section>
+          <section className="content_section">
+            {currentTab === 'list_tab' ? (
+              <ListView diaryData={diaryData} handleDelete={handleDelete} />
+            ) : (
+              <GalleryView diaryData={diaryData} />
+            )}
+          </section>
+          <div className="top_btn"></div>
+        </div>
+        <div className="write_btn_wrap">
+          <Link to="/diary/write" className="write_btn"></Link>
+        </div>
+      </main>
+      <Footer />
+    </>
   );
 };
